+++ src/components/folderPage/FolderCategoryItem.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from "react";
 import styled from "styled-components";
-import { useFolder } from "../context/FolderContext";
+import { useFolder } from "../../context/FolderContext";
 
 const DefaultFolder = styled.button`
   padding: 8px 12px 8px 12px;
