+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
-import { FolderProvider } from "./folder/context/FolderContext";
+import { FolderProvider } from "../context/FolderContext";
 import Header from "../components/header/Header";
 
 export default function App({ Component, pageProps }: AppProps) {
