+++ src/pages/folder/index.tsx
@@ -1,12 +1,12 @@
-import Banner from "./components/Banner";
+import Banner from "../../components/folderPage/Banner";
 import SearchBar from "@/src/components/SearchBar";
-import FolderCategoryList from "./components/FolderCategoryList";
-import FolderTitle from "./components/FolderTitle";
-import FolderCardList from "./components/FolderCardList";
+import FolderCategoryList from "../../components/folderPage/FolderCategoryList";
+import FolderTitle from "../../components/folderPage/FolderTitle";
+import FolderCardList from "../../components/folderPage/FolderCardList";
 import Footer from "@/src/components/Footer";
-import FolderAddButton from "./components/FolderAddButton";
-import FolderOption from "./components/FolderOption";
-import SearchResultTitle from "./components/SearchResultTitle";
+import FolderAddButton from "../../components/folderPage/FolderAddButton";
+import FolderOption from "../../components/folderPage/FolderOption";
+import SearchResultTitle from "../../components/folderPage/SearchResultTitle";
 import styled from "styled-components";
 
 const Section = styled.section`
